Memperbaiki dropdown nama Unit & Biaya HPP

+++ database/migrations/2023_12_04_020323_create_biaya_hpps_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->foreignId('unit_id')->constrained('unit');
             $table->string('jenis_biaya');
-            $table->float('biaya_per_gram');
+            $table->integer('biaya_per_gram');
             $table->integer('status')->default('1');
             $table->timestamps();
         });
